Fix numbers being cut out from messages
parseFloat() is used to parse the tokenized string and if it does not return a NaN then it considers the token to be a number. This funnily returns non-NAN for numbers with multiple ".", such as `[connect 192.168.7.2(`

Not sure why it even bothers distinguishing between numbers and strings, but maybe I am missing something.
Anyhow, I removed the parseFloat() call and now it works nicely.